Update custom-server example
